+++ R/install_tika.R
@@ -11,10 +11,9 @@
 #' @param url Optional url to a particular location of the tika app. Setting this to any character string overrides downloading from random mirrors.
 #'
 #' @return Logical if the installation was successful.
-#' @examples
-#' \donttest{
+#' @examplesIf interactive()
 #' install_tika()
-#' }
+#' 
 #' @section Details:
 #' The default settings of \code{install_tika()} should typically be left as they are.
 #'
@@ -27,7 +26,7 @@
 #' You can also enter a value for \code{url} directly to override this.
 #'
 #' It will download into a directory determined
-#' by the \code{rappdirs::user_data_dir()} function,
+#' by \code{tools::R_user_dir("rtika", which = "data")},
 #' specific to the operating system.
 #'
 #' If \code{tika()} is stopping with an error compalining about the \code{jar},
@@ -38,12 +37,11 @@
 #' and want to remove the Tika App \code{jar} also,
 #' run:
 #'
-#'  \code{unlink(rappdirs::user_data_dir('rtika'), recursive = TRUE)}
+#'  \code{unlink(tools::R_user_dir("rtika", which = "data"), recursive = TRUE)}
 #'
 #' Alternately, navigate to the install folder and delete it manually.
-#' It is the file path returned by \code{rappdirs::user_data_dir('rtika')}.
-#' The path is OS specific, and explained here:
-#' https://github.com/r-lib/rappdirs .
+#' It is the file path returned by \code{tools::R_user_dir("rtika", which = "data")}.
+#' The path is OS specific.
 #' 
 #' @section Distribution:
 #' Tika is distributed under the Apache License Version 2.0,
@@ -69,7 +67,7 @@ install_tika <- function(version = "2.0.0",
      # Get user directory  -------------------
      user_data_dir <-
     normalizePath(
-      rappdirs::user_data_dir("rtika"),
+      R_user_dir("rtika", which = "data"),
       mustWork = FALSE
     ) # tools::R_user_dir works on R > 4